fix 64-bit compile and segfault
git-svn-id: https://gambatte.svn.sourceforge.net/svnroot/gambatte@48 9dfb2916-2d38-0410-aef4-c5fe6c9ffc24

+++ gambatte_qt/src/videodialog.cpp
@@ -229,7 +229,7 @@ void VideoDialog::fillFullResSelector(const QSize &minimumRes) {
 		const int vres = resVector[i].h;
 		
 		if (hres >= minimumRes.width() && vres >= minimumRes.height())
-			fullResSelector->addItem(QString::number(hres) + QString("x") + QString::number(vres), i);
+			fullResSelector->addItem(QString::number(hres) + QString("x") + QString::number(vres), (uint)i);
 	}
 }
